[hotfix][runtime] Extract SharedStateRegistry interface
