+++ flink-runtime/src/test/java/org/apache/flink/runtime/checkpoint/ZooKeeperCompletedCheckpointStoreITCase.java
@@ -23,6 +23,7 @@ import org.apache.flink.api.common.JobStatus;
 import org.apache.flink.api.common.time.Deadline;
 import org.apache.flink.runtime.state.RetrievableStateHandle;
 import org.apache.flink.runtime.state.SharedStateRegistry;
+import org.apache.flink.runtime.state.SharedStateRegistryImpl;
 import org.apache.flink.runtime.testutils.CommonTestUtils;
 import org.apache.flink.runtime.util.ZooKeeperUtils;
 import org.apache.flink.runtime.zookeeper.ZooKeeperStateHandleStore;
@@ -104,7 +105,7 @@ public class ZooKeeperCompletedCheckpointStoreITCase extends CompletedCheckpoint
     @Test
     public void testRecover() throws Exception {
 
-        SharedStateRegistry sharedStateRegistry = new SharedStateRegistry();
+        SharedStateRegistry sharedStateRegistry = new SharedStateRegistryImpl();
         CompletedCheckpointStore checkpoints = createRecoveredCompletedCheckpointStore(3);
 
         TestCompletedCheckpoint[] expected =
@@ -132,7 +133,7 @@ public class ZooKeeperCompletedCheckpointStoreITCase extends CompletedCheckpoint
 
         // Recover
         sharedStateRegistry.close();
-        sharedStateRegistry = new SharedStateRegistry();
+        sharedStateRegistry = new SharedStateRegistryImpl();
 
         assertEquals(3, ZOOKEEPER.getClient().getChildren().forPath(CHECKPOINT_PATH).size());
         assertEquals(3, checkpoints.getNumberOfRetainedCheckpoints());
@@ -161,7 +162,7 @@ public class ZooKeeperCompletedCheckpointStoreITCase extends CompletedCheckpoint
     public void testShutdownDiscardsCheckpoints() throws Exception {
         CuratorFramework client = ZOOKEEPER.getClient();
 
-        SharedStateRegistry sharedStateRegistry = new SharedStateRegistry();
+        SharedStateRegistry sharedStateRegistry = new SharedStateRegistryImpl();
         CompletedCheckpointStore store = createRecoveredCompletedCheckpointStore(1);
         TestCompletedCheckpoint checkpoint = createCheckpoint(0, sharedStateRegistry);
 
@@ -197,7 +198,7 @@ public class ZooKeeperCompletedCheckpointStoreITCase extends CompletedCheckpoint
     public void testSuspendKeepsCheckpoints() throws Exception {
         CuratorFramework client = ZOOKEEPER.getClient();
 
-        SharedStateRegistry sharedStateRegistry = new SharedStateRegistry();
+        SharedStateRegistry sharedStateRegistry = new SharedStateRegistryImpl();
         CompletedCheckpointStore store = createRecoveredCompletedCheckpointStore(1);
         TestCompletedCheckpoint checkpoint = createCheckpoint(0, sharedStateRegistry);
 
@@ -238,7 +239,7 @@ public class ZooKeeperCompletedCheckpointStoreITCase extends CompletedCheckpoint
     @Test
     public void testLatestCheckpointRecovery() throws Exception {
         final int numCheckpoints = 3;
-        SharedStateRegistry sharedStateRegistry = new SharedStateRegistry();
+        SharedStateRegistry sharedStateRegistry = new SharedStateRegistryImpl();
         CompletedCheckpointStore checkpointStore =
                 createRecoveredCompletedCheckpointStore(numCheckpoints);
         List<CompletedCheckpoint> checkpoints = new ArrayList<>(numCheckpoints);
@@ -273,7 +274,7 @@ public class ZooKeeperCompletedCheckpointStoreITCase extends CompletedCheckpoint
         final CompletedCheckpointStore zkCheckpointStore1 =
                 createRecoveredCompletedCheckpointStore(numberOfCheckpoints);
 
-        SharedStateRegistry sharedStateRegistry = new SharedStateRegistry();
+        SharedStateRegistry sharedStateRegistry = new SharedStateRegistryImpl();
 
         TestCompletedCheckpoint completedCheckpoint = createCheckpoint(1, sharedStateRegistry);
 
@@ -283,7 +284,7 @@ public class ZooKeeperCompletedCheckpointStoreITCase extends CompletedCheckpoint
 
         // recover the checkpoint by a different checkpoint store
         sharedStateRegistry.close();
-        sharedStateRegistry = new SharedStateRegistry();
+        sharedStateRegistry = new SharedStateRegistryImpl();
         final CompletedCheckpointStore zkCheckpointStore2 =
                 createRecoveredCompletedCheckpointStore(numberOfCheckpoints);
